+++ CHANGELOG.md
@@ -2,7 +2,9 @@
 ## 2.5 branch
 ### 2.5.0
 * now it uses the `mirko-pagliai/php-tools` package. This also replaces
-    `mirko-pagliai/reflection`.
+    `mirko-pagliai/reflection`;
+* removed `BackupTrait::getClassShortName()` method. The
+    `get_class_short_name()` global function will be used instead.
 
 ## 2.4 branch
 ### 2.4.0